Changes in preparation for version 1.3 release.
Added symlog_linear_plot and fill_between_curves examples.
Fixed a bug where the upper axis limit would be autoselected as the data limit instead of the major tick mark on log scaled axes.
Added the filled regions from the new fill_between_x_curves and fill_between_y_curves to the list of objects that are used to autoselect axis limits.
Fixed a bug in plot_box_and_whiskers that messed up the tick marks on the x-axis.
Converted any string in the examples and tests with a '\' to a raw string to avoid annoying warning messages from python.
Updated the version history.